+++ lib/eve_online/esi/corporation.rb
@@ -7,8 +7,6 @@ module EveOnline
     class Corporation < Base
       extend Forwardable
 
-      include ResponseWithEtag
-
       API_ENDPOINT = 'https://esi.evetech.net/v4/corporations/%<corporation_id>s/?datasource=%<datasource>s'
 
       attr_reader :corporation_id
@@ -23,10 +21,10 @@ module EveOnline
       def_delegators :model, :as_json, :alliance_id, :ceo_id, :creator_id,
                      :date_founded, :description, :faction_id,
                      :home_station_id, :member_count, :name, :shares,
-                     :tax_rate, :ticker, :corporation_url, :etag
+                     :tax_rate, :ticker, :corporation_url
 
       def model
-        @model ||= Models::Corporation.new(response_with_etag)
+        @model ||= Models::Corporation.new(response)
       end
 
       def scope; end