+++ lib/eve_online/esi/models/alliance.rb
@@ -4,8 +4,6 @@ module EveOnline
   module ESI
     module Models
       class Alliance < Base
-        include ModelWithEtag
-
         def as_json
           {
             creator_corporation_id: creator_corporation_id,
@@ -14,8 +12,7 @@ module EveOnline
             executor_corporation_id: executor_corporation_id,
             faction_id: faction_id,
             name: name,
-            ticker: ticker,
-            etag: etag
+            ticker: ticker
           }
         end