+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 ## master
 
 * Add `EveOnline::ESI::AllianceCorporations#etag`
+* Remove `EveOnline::ESI::Models::ModelWithEtag`
+* Remove `EveOnline::ESI::ResponseWithEtag`
+* Remove `EveOnline::ESI::Models::Alliance#etag`
+* Remove `EveOnline::ESI::Models::Corporation#etag`
+* Add exception `EveOnline::Exceptions::NotModified` for case when body not present for 304 request
+* Remove `EveOnline::ESI::Models::Star#etag`
+* Remove `EveOnline::ESI::Models::Stargate#etag`
+* Add `EveOnline::ESI::Base#etag`
+* `EveOnline::ESI::Base#content` raise `EveOnline::Exceptions::NotModified` when body not present for 304 requests
 
 ## v0.21.0